Moving stuff around to create a (simple) module
