+++ src/ddos_dissector/exceptions/UnsupportedFileTypeError.py
